Move Service and HttpService into core.
