+++ examples/blaze/src/main/scala/com/example/http4s/blaze/BlazeMetricsExample.scala
@@ -5,7 +5,8 @@ package com.example.http4s.blaze
 import java.util.concurrent.TimeUnit
 
 import com.example.http4s.ExampleService
-import org.http4s.server._
+import org.http4s._
+import org.http4s.server.Router
 import org.http4s.server.blaze.BlazeBuilder
 import org.http4s.server.middleware.Metrics
 import org.http4s.dsl._
