+++ examples/src/main/scala/com/example/http4s/site/HelloBetterWorld.scala
@@ -1,8 +1,8 @@
 package com.example.http4s
 package site
 
+import org.http4s._
 import org.http4s.dsl._
-import org.http4s.server.{HttpService, Service}
 
 object HelloBetterWorld {
   /// code_ref: service
