+++ server/src/main/scala/org/http4s/server/package.scala
@@ -6,91 +6,8 @@ import scalaz.syntax.kleisli._
 
 package object server {
 
-  /**
-   * A Service wraps a function of request type [[A]] to a Task that runs
-   * to esponse type [[B]].  By wrapping the `Service`, we can compose them
-   * using Kleisli operations.
-   */
-  type Service[A, B] = Kleisli[Task, A, B]
 
-  object Service {
-    /**
-     * Lifts an unwrapped function that returns a Task into a [[Service]].
-     *
-     * @see [[HttpService.apply]]
-     */
-    def lift[A, B](f: A => Task[B]): Service[A, B] = Kleisli.kleisli(f)
 
-    /**
-      * Lifts a Task into a [[Service]].
-      *
-      */
-    def const[A, B](b: => Task[B]): Service[A, B] = b.liftKleisli
-
-    /**
-      *  Lifts a value into a [[Service]].
-      *
-      */
-    def constVal[A, B](b: => B): Service[A, B] = Task.now(b).liftKleisli
-
-    /**
-      * Allows Service chainig through an implicit [[Fallthrough]] instance.
-      *
-      */
-    def withFallback[A, B : Fallthrough](fallback: Service[A, B])(service: Service[A, B]): Service[A, B] =
-      service.flatMap(resp => Fallthrough[B].fallthrough(resp, fallback))
-
-  }
-
-  /**
-   * A [[Service]] that produces a Task to compute a [[Response]] from a
-   * [[Request]].  An HttpService can be run on any supported http4s
-   * server backend, such as Blaze, Jetty, or Tomcat.
-   */
-  type HttpService = Service[Request, Response]
-
-  /**
-    * There are 4 HttpService constructors:
-    * <ul>
-    *  <li>(Request => Task[Response]) => HttpService</li>
-    *  <li>PartialFunction[Request, Task[Response]] => HttpService</li>
-    *  <li>(PartialFunction[Request, Task[Response]], HttpService) => HttpService</li>
-    *  <li>(PartialFunction[Request, Task[Response]], Task[Response]) => HttpService</li>
-    * </ul>
-    */
-  object HttpService {
-
-    /** Alternative application which lifts a partial function to an `HttpService`,
-      * answering with a [[Response]] with status [[Status.NotFound]] for any requests
-      * where the function is undefined.
-      */
-    def apply(pf: PartialFunction[Request, Task[Response]], default: HttpService = empty): HttpService =
-      Service.lift(req => pf.applyOrElse(req, default))
-
-    /** Alternative application  which lifts a partial function to an `HttpService`,
-      * answering with a [[Response]] as supplied by the default argument.
-      */
-    def apply(pf: PartialFunction[Request, Task[Response]], default: Task[Response]): HttpService =
-      Service.lift(req => pf.applyOrElse(req, (_: Request) => default))
-
-    /**
-      * Lifts a (total) function to an `HttpService`. The function is expected to handle
-      * ALL requests it is given.
-      */
-    def lift(f: Request => Task[Response]): HttpService = Service.lift(f)
-
-    /** The default 'Not Found' response used when lifting a partial function
-      * to a [[HttpService]] or general 'not handled' results.
-      * 
-      * This particular instance is tagged with an so that it interacts appropriately
-      * attribute to play well with the default [[Fallthrough]] behavior.
-      */
-    val notFound: Task[Response] = Task.now(Response(Status.NotFound)
-                                               .withAttribute(Fallthrough.fallthroughKey, ())
-                                               .withBody("404 Not Found.").run)
-    
-    val empty   : HttpService    = Service.const(notFound)
-  }
 
   /**
    * A middleware is a function of one [[Service]] to another, possibly of a
