+++ server/src/test/scala/org/http4s/server/middleware/UriTranslationSpec.scala
@@ -1,9 +1,9 @@
-package org.http4s.server.middleware
+package org.http4s
+package server
+package middleware
 
 import java.nio.charset.StandardCharsets
 
-import org.http4s.{Http4sSpec, Uri, Request, Response}
-import org.http4s.server.HttpService
 import org.http4s.Status._
 
 import org.specs2.mutable.Specification
